app/requirements-win.txt
fastapi
uvicorn[standard]
tensorflow-cpu
pillow
numpy
python-multipart
